HPCC-14462 Fix subfile removal when only one file
If in ECL Watch you open a superfile which contains a single subfile,
check the box and click on remove then though the file is removed and the
file size in the Summary area is updated, the sub-file is still shown. It
remains even if 'Refresh' is clicked, or the file tab is closed and re-opened.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>
Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>