HPCC-14416 Create compressed empty files when writing larger file
When a file is output to a larger cluster, the number of parts is
padded to match the destination cluster.
If the target file is compressed the empty physical parts need to
be properly constructed, so that they have a compressed format
header. Prior to this fix, it used to create a 0 length file part,
which was fine if uncompressed, but caused read back issues if
the file was compressed.

Most of the diff in this commit, is moving functionality that
used to be in thmfilemanager.cpp to the diskwrite activity (the
only place it is needed), so that it can read needed info from the
activity (compression type, recordsize)

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>