HPCC-15513 Fix missing Build Version in ZAP report
In the existing WsWorkunits.WUCreateZAPInfo, both ESP Build Version
and ESP Network Address are added to ZAP report based on the values
inside the WUCreateZAPInfo request (which are forwarded by Create
ZAP Report form). When an ECLWU ZAP report is not created using the
form, both ESP Build Version and ESP Network Address are empty in
the report. In this fix, the code is added to read those values
into the report.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>