HPCC-15666 Fix multipart index regressions
1) The current keymanager in use index, was not being reset
for the next execution of a child query.
2) Index read with a keyed limit, gets the key count upfront.
The rest of the index read uses a key merger. The merger doesn't
handle getCount properly and uses a key manager for each key part
to aggregate the count. When resetting the key manager, it needs
to avoid resetting the unused key merger manager.
3) A few problems with eoi in derived activities, e.g. IndexCount
4) IndexCount getCount() usage fix, can't use checkCount

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>