HPCC-15660 Can't pass stored values to EMBED options in Thor
Rework the code generated for EMBED options so that it is calculated outside
of the generated user function and passed in. This will mean that it can be
calculated just once (on the master, in Thor) and serialized to the slaves.

Don't create the options parameter for embedded C++, and suppress the hidden
options parameter from regenerated ECL.