HPCC-15712 Spurious warnings from generated code when using Clang
This is a bit of a hack, but there is another Jira open for a proper fix
(introducing a new CLang compiler type). And it makes my life a lot easier in
the meantime.

Also has a workaround for a gcc compiler bug that makes the warnings
referenced in HPCC-15713 fatal errors on some platforms. HPCC-15713 should
remain open (and we should move towards warning-free generated code).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>