HPCC-13041 No error message when uploading a file
Whenever a user tries to upload a file to a directory that they do no have permissions they do not get a failed error or anything of the sort. The exception is thrown by the client side and we are capturing it and testing against it.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>