HPCC-15840 Master not passing aux rowIf in join.
If no keyserializer was generated, then the standard row
serializer was used. The slaves correctly setup based on the
primary partitioning side, the master however didn't pass the
primary row interface and based on the other side.
And hence caused serialization problems.

NB: This was only an issue if no keySerializer generated in a
global join.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>