HPCC-15873 Add FileUploadAccess to default permissions
When ESP SMC service is started, the service reads feature
permission settings from esp.xml and add the permission to
LDAP server if the permission is not in the LDAP server.
In this fix, FileUploadAccess is added to environment
configuration files which creates the FileUploadAccess
permission setting in esp.xml.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>