HPCC-15829 Race condition spilling result stream after read
A small window existed where a result stream could have been
fully read, then spilt, then nextRow() called again.
Under those circumstances, the spill would not setup the stream
read offset and the result stream was not marked as having been
completely ready (at eos). As a result an attempt to read the
spill stream was made, but an assert was hit.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>