HPCC-15953 Ordering fix for Local wuid write with EXTEND
If OUTPUT is used in a child query (e.g. via APPLY) and there
are multiple outputs writnig to the same named workunit result
with EXTEND, then the output order was not deterministic.

Fix is to avoid acknowledging data packets read from slave
workunit writes until data flushed to workunit.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>