HPCC-15950 Race condition in firstn inside child query
The race causes a global firstn inside a child query to
potentially hit an assert(read <= limit) in the master.
The bug was caused by the firstn slave signalling a semaphore
twice, once when it hit the end of input via lookahead and once
via stop(). The result was that subsequent executions of the
same activity would spuriously fire early and could report to the
master that they were going to read more than the limit.

Also removed spurious code in onInputFinished when count was 0

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>