HPCC-16051 Fix problems with child act. initialization
Activities with variant initialization, e.g. dynamic filenames
were not being initialized correctly if in loop or child queries.
In some cases it could cause a crash (this JIRA was an incident of
that), in others it would fail to re-initialize based on the
context. Also changed so that only the child activities that
need initialization, request it, which in many cases means none.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>