HPCC-16120 Fix type causing regression in init_thor script
The section of code for kill_slaves relies on the uslaves file
being present.  Yet in the frunssh call we use the slaves file
for our ip list, unlike every other frunssh call where we rely on
the uslaves file.  This was causing issues where the uslaves file
would be present, but a previous failure to connect to the slave
machines would erase the slaves file (as was intended), and cause
the startup to always fail.

The fix is to use the uslaves file, as was originally intended.
If there is a failure to connect to the slave machines to stop the
slaves, we want the uslaves file to remain in the instancedir so
we can try again on the next startup attempt.

Signed-off-by: Michael Gardner <michael.gardner@lexisnexis.com>