HPCC-15945 Thor slave hang in countproject
If a slave countproject calls stop() before nextRow()
(ie if one slave hits CHOSEN limit others will be told to stop)
a deadlock could occur.  This change moves PARENT::stop() call to
after check for first in stop() so that signalNext() can signal
the semaphore inInputFinished() is waiting for.  Thus PARENT::stop()
does not deadlock waiting for onInputFinished to end.

Signed-off-by: Mark Kelly <mark.kelly@lexisnexis.com>