HPCC-16130 Change default timeout on ok_to_send response
If packets get lost in the ok_to_send conversation, then waiting 5 seconds
before resending is too long if the entire master->slave conversation is going
to time out in 2 seconds (as it does by default...)

Seems more sensible to default to timing out the ok_to_send in slightly less
than the master-slave.

Also, change the timeout to be specified in milliseconds rather than seconds
for consistency and granularity.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>