HPCC-16132 Missing foreign subfiles cause problems to superfile ops.
If a foreign subfile is deleted, the superfile removed the entry
(for the purposes of ignoring it) from the meta info temporarilty
whilst the superfile was in use.
Unfortunately that caused problems if the use context was
manipulating the superfile, e.g. adding/removing subfiles.

It also prevented that missing superfile being removed from the
superfile for the same reason.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>