HPCC-16266 Fix issue with SOAP DATASET/ACTION inside child query
IWSCHelper helper needs to be created in start() after onStart()
call, as it may be dependent on start context.

Also SOAP dataset call was not calling IWSCHelper::waitUntilDone()
in the activities stop() method (the other SOAP variants were).
This could (race) cause 'Thread already started!' messages when
executed in a child query.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>