HPCC-14493 Cross domaing scripting of ESP API
Adds CORS interoperability to ESP, but only for interoperability with
cross domain browser calls.  We do not treat the browser as a trusted
entity or rely on it to make security decisions for us.

Since we implement general purpose APIs and not just scripted pages
we need to be diligent and secure for every request whether it comes
from a cross domain browser, a REST or SOAP client, or any other source.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>