HPCC-16355 Fix compile error in dalift
In HPCC-16309, the system/security/shared is added into the include_directories
section of the dali/ft/ftslave.cmake. But, the ftslave.cmake is not used by
clienttools build which causes a compile error in dalift for the clienttools
build: Cannot open include file: 'seclib.hpp': No such file or directory. In
this fix, the system/security/shared is added into the include_directories
section of the dali/ft/dalift.cmake.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>

section of the dali/ft/ommit message for your changes. Lines starting