HPCC-16179 Redundant LDAP server not used
The LDAP connection pool incorrectly creates connections to all LDAP servers
listed in the configuration. This is bad because a resource could be created
on one LDAP, followed by attempts to reference it on another LDAP.
The correct behavior is to use the same one, and
if it fails move on to the next one. This PR replaces the old CLoadBalancer
class with a CHostManager, which implements the correct behavior

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>