HPCC-16634 Disabled selections block log access
When incorporating preflight in the advanced form area some selections were disabled to keep users from selecting options that were not applicable to submit to preflight results. As a result, the logs tabs were not being highlighted for some of the components.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>