HPCC-16751 Prevent potential join crash on stop()
If an upstream activity called Join::stop() twice (which it shouldn't), it
caused a crash.
The particular case where this was spotted, was where a lightweight selfjoin
called it's input stop() twice.
It had to be a lightweight variety, where the code called through to the
input stop() directly and then again in the slave activity base class.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>