HPCC-16841 repository not populating in ECL IDE
LdapServer security Manager calls to LDAP time out after 20 seconds. Calls
to enumerate the repository are currently taking up to 21 seconds, and likely
more in other cases. This PR sets the default timeout to 60 seconds, and adds
and LDAPServer configuration option for clients to set it to any number of
seconds

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>