HPCC-16952 Loading eclblas plugin on OSX causes fork() to core
This appears to be a known issue with openBlas, fixed in more recent versions.

As a quick workaround, and since it's a desirable change anyway, move the
eclblas dll out of plugins and into libs - we can and should do this for any
plugin for which there is a .ecllib proxy.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>