HPCC-16839 Avoid issue when asyncBackup=false and backup node down
If asyncBackup was off and there was a failure to communicate with
the backupnode, it caused the block being committed to be retained
in memory and rewritten to the primary transaction delta log.
Consequently, when a coalesce or a fatal dali restart occurred
that relied on the transaction log, errors loading these duplicate
delta transactions would occur.

Also ensure delta transaction and store are replicated to the
backup node at startup.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>