HPCC-16972 Cassandra workunit store not handling deletes properly
Deleting and then adding some exceptions within a single workunit commit would
end up losing the newly added ones - this is because of how Cassandra batches
work when rows are both deleted and updated within the same batch. I doubt
this would cause any issues in practice outside of test suite code.

Also, removing certain information from a workunit - specifically the file
associations but there may be other instances - would result in this
information not being properly committed. We need to bind NULL to any columns
that we want cleared in an update.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>