HPCC-16906 Add a new plugin for embedded Python3 code
Copied from the Python2 plugin, with modifications to reflect the changed C
API.

Note that libpython3.x exports many of the same symbols as libpython2.x, whcih
can lead to issues if both are loaded. In particular, I had to change the
visibility rules when loading plugins, and I also had to refine the code for
working around Centos distro issue with improperly linked Python extensions.
We now only apply the workaround if py3embed is not in use - people using
centos and buggy python distros will need to pick one or the other version of
pyembed, but not both.

Also discovered that one of the existing tests was working by luck (since
python sets are unordered), and stopped working on (at least some) python 3
installations.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>