HPCC-17144 Strange code in getScopePermissions
getScopePermissions is allowing callers to pass in an invalid auditflags
argument which callers use to determine whether or not permissions are
available (security enabled).  Callers use this hack to first determine
if they should perform security checks.  This PR removes that peculiar
behavior and exposes the once private "permissionsavail" flag, which allows
callers to perform this check in a more intuitive manner

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>