HPCC-17231 Add option to control sort spill compressed block size
The compressed spilt streams of sort, are read by a merger, each
compressed stream reads a block at a time and decompressed it.
Depending on the compression ratio and the number of spills, this
can use a very large amount of heap memory.
This option allows the written stream to use small block sizes to
mitigate the read/decompressed issues.

NB: This PR also fixes ICompressedFileIO::setBlockSize, which was
being ignored, because it could not be called early enough/the
compressor had already been setup with the default size.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>