HPCC-13443 Avoid spurious 'Invalid CPTStack detected'/stack tracing
A disconnecting connection whose root node has been deleted by
someone else caused 'Invalid CPTStack detected' and a stack trace
to be produced. This was in code that was calculating the absolute
path to be recorded in a saved delta, however in this case it
wasn't needed/used anyway.
Detect orphaned node and avoid calling the code.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>