HPCC-17280 Embedded R code uses persistent, shared environment
Made a change to work around the fact that the two-parameter form of _new_env
does not appear to be present in older versions of Rcpp.

We could take the view that updating the version of Rcpp on the build machines
is preferable in whcih case this commit should be removed.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>