HPCC-17314 Fix excessive splitter tracing bug
When a splitter output is stalled, it is recorded so that it
can be signalled to be unblocked, it is also used to trace all
stalled outputs if stalled for a longer period.
When unstalled the index array which records which outputs are
stalled was not being cleared.
Meaning that if splitter outputs regularly stalled/released,
the index array would grow and the tracing would spuriously
trace many duplicate lines mentioning the same output.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>