HPCC-17331 Prevent Smartjoin crash when failing over to local
The root of the issue is still unknown, this change guards against
what was seen to be happening (a non-local row being processed at
a late stage during local failover) and adds a little more tracing
to help diagnose.

By default the non-local row will still cause the query to fail
(with an error and log the row), so that we can see and have any
other incident of this issue reported.
However, the user can workarond the issue by enabling an option
to suppress the error.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>