HPCC-17105 Default created superfiles in PromoteSuperFile to interleaved
To match elsewhere.
To not be interleaved meant superfiles would split parts sequentially,
therefore instead of reading all part 1's on node 1, it would read
part 1-M on node 1, part M+1-N on node 2 etc. And consequently where
the cluster was local to the logical files, do a lot of unnecessary
remote reading.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>