HPCC-17181 Refactor aggregate HT to use Roxiemem+merge sort locals
Previously hash aggregates, were creating a rowbuilder object per new
row on the heap. With a large number of unique values, this created a
lot of heap objects that would be randomly accessed as the hash table
was used. This was particularly expensive (very variable time) during
the merge phase. In most extreme cases it was 60x slower than
fastest run.

This change instead pre-sorts the local hash tables, before performing
a distribute merge. This has the secondary advantage of being able to
stream out the aggregated values on each node, instead of having to
wait for the whole global aggregate to have finished.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>