HPCC-17040 Follow-up errors could be returned before primary cause
An exception in caught in the master graph that caused a query to
abort, could lead to follow-up exception that would be reported
back to the job before the primary exception was reported.

Special handling of in master of abort type exception was leaving
a window where this could happen, change to let job handler report
and handle abort.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>