HPCC-17464 A sink with an UPDATE, could cause balanced splitter stall
If a OUTPUT/BUILD,UPDATE was pruned/never executed and there was an
upstream balanced splitter, the splitter was unaware of the
non-executing arm and would deadlock when it wasn't pulled once the
balanced splitter buffer was full.

In the case of an unbalanced splitter, it would not stall, but the
splitter input would never be stopped, which cause other issues.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>