HPCC-17650 Definition dropdown has stale data
The definition drop down can have stale data due to the initialization happeningonce, which is when the call is made. If the user is adding a definition somewhere else they will need to reload the page to get the new definition list. I have modified the initialization to happen everytime someone opens the dialog to always have the latest values.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>