HPCC-17998 Ensure dali session tree for ESP using old env xml
We want to make sure that ESP is running in session management
even if old environment.xml is used. ESP should create the session
tree in dali for that situation.

Also should not setup dali session tree for an ESP binding if no
session authentication is required for that binding.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>