HPCC-17826 Set IsCompressed for key file in WsDFU.DFUInfo
For key file, the IsCompressed flag in WsDFU.DFUQuery response is
set to true. But, the flag is not set in the existing WsDFU.DFUInfo
response. The problem is fixed in this PR. This PR also uses file
size as the CompressedFileSize for key file since a key file is
compressed.