HPCC-17967 Prevent XREF scanning duplicate nodes and misreporting
XREF was treating every entry in the cluster groups as a separate
node to scan. As a consequence it would scan the same node
directories N times, on clusters with slavesPerNode>1
As a side effect, the overscanning caused some reports to be
inaccurate, e.g. caused number of orphan file parts to be N times
larger than they actually were.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>