HPCC-18269 Avoid repeatedly processing long CSV line text.
The CSV reading code that processed long lines had a bug in it
that caused it to re-process text too many times when trying to
read to the end of the line.
Instead of asking for more than had been returned by the last
read call, it asked for double what it had previously asked for.
Which was usually a lot smaller than what had been returned.
Consequently, the loop would spin for a while re-processing the
partial CSV line until the amount asked for actually exceeded the
amount available.

Also commoned up code into a stream reading method.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>