HPCC-18296 WsWorkunits should ignore debug values that begin with '-'
Some eclcc options passed through eclcmd result in debug options that
begin with '-', these should not be added to the workunit as hash
options.  Fixing in the webservice acts as a good defensive measure
against bad debug option parameters.

Should consider changing how the eclcmd handles the two types of options
in the future.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>