HPCC-18103 Keep compressed icon consistent
Whenever a user goes into a super file / logical file that is compressed the only thing that appeared with compressed size and percent compressed (which you have to scroll down for). There was no initial icon that did indeed show that the file was compressed. Adding an icon at the top of the details page to show this information.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>