HPCC-18354 Unable to download Thor slave logs
When trying to download a log file from a thor, the ecl complains it cannot read the file, although it can display the content of the file. The same happens with any log file, this one is just one example. The netaddress was not being passed in. I have added it to the Name parameter.

Signed-off by: Miguel Vazquez <miguel.vazquez@lexisnexis.com>