HPCC-11822 Add new code to access archived workunits with paging
In saarch.cpp, WUiterateSorted() is added to retrieve archived workunits
sorted by WUID. The new code is based on sortDirectory(). The beforeWU/
afterWU and/or StartDate/EndDate are used to skip the directories which
store WUs not in the date range. The beforeWU/afterWU is also used to
skip the WUs which are not in the time range. The exisitng filters in
the WUiterate() are supported by WUiterateSorted().

WUiterateSorted() is used by WsWorkunits.WUQuery to access archived
workunits with paging. An ESP client may ask for next page by specifing
afterWU or ask for previous page by specifing beforeWU.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>