HPCC-11822 Add new code to access archived workunits with paging
In saarch.cpp, WUiterate() is refactored to retrieve archived workunits
sorted by WUID. The new code is based on IFile::directoryFilesSorted().
The beforeWU/afterWU and/or StartDate/EndDate are used to skip the
directories which store WUs not in the date range. The beforeWU/afterWU
is also used to skip the WUs which are not in the time range.

The refactored WUiterate() is used by WsWorkunits.WUQuery to access
archived workunits with paging. An ESP client may ask for next page by
specifing afterWU or ask for previous page by specifing beforeWU.

Revise based on code review:
1. Merging WUiterateSorted into the existing WUiterate;
2. adding code to support SCA_WORKUNIT_SERVICES_GET, SCA_LISTDT,
and SCA_GET;
3. add sorted directory iteration into jlib.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>