HPCC-18596 Avoid patching workunit XML patching
Exporting a workunit to XML invovled serializing it to XML
and spotting parts which were sensitive and removing them on
the fly. Followed by a later stage which inserted another piece
of XML into the XML text block.

Instead get a copy of the workunit tree, prune the unwanted into,
add the extra info (in the form of their original trees) and
serialize the whole cleaned workunit in one go.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>