HPCC-18626 Ensure value subscriber notification for replaced trees.
When a subscriber subscribes to a tree which is replaced by
another, the value in the notificaiton was lost.
A replaced tree generates a delete event and a new value
event, in that order, ensure that the new value gets
through, by scanning the changes per node in reverse order.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>