HPCC-18658 Refactor ESP sessionID and digital signature data types
Currently ESP stores a user's digital signature and ESP Session ID in
MemoryBuffers. The ESP Session ID is really just an unsigned value,
and should be stored and returned as such. The digital signature will
be a base 64 encoded string, and should be passed as const char *.
This PR makes those changes

HPCC-18658 Refactor ESP sessionID and digital signature data types

Currently ESP stores a user's digital signature and ESP Session ID in
MemoryBuffers. The ESP Session ID is really just an unsigned value,
and should be stored and returned as such. The digital signature will
be a base 64 encoded string, and should be passed as const char *.
This PR implements those changes

Signed-off-by: Russ Whitehead <william.whitehead@lexisnexis.com>