HPCC-18672 Access fileposition from an index row instead of special casing
The current indexes store a fileposition with each index entry, but stored
before the row.  Previously this value was accessed by adding extra
parameters to functions and functions to interfaces.

This commit moves the fileposition into the logical row by appending it.  This
means it can be treated in the same way as all the other fields in a key.
The code generator now retrieves the fileposition from the row, and the layout
mapping code is simplified to remove the special casing.

The commit ensures CRCs for index files stay identical to avoid backward
compatibility problems.

A future commit will modify the activity helpers and other code that can now be
simplified.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>