HPCC-18725 Optimize dafilesrv size() and getInfo() operations
size() and getInfo() are very expensive if the files they
operate on are inside directories with lots of files.
This is because both operations use the remote command getDir
to retrieve 1 entry and despite it only serializing 1 item back
it is very expensive on the server side at the OS level and in
dafilesrv to scan large directories.

Change to spot directory request involving trivial case (as used
by size() and getInfo()) where there is no recursion and no
wildcards. In these cases a simple direct single file request
can be made and the same information can be serialized back.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>