HPCC-18738 Pass current user creds, if no foreign creds available
Ensure that the current users credentials are supplied to DFS
lookup calls on a remote Dali operatation if foreign credentials
have not been explicitly supplied.

Without this, the default was to pass no credentials, meaning a
system with authentication on that disabled access to the default
user would be deny all foreign remote operations, e.g. a
File.Std.Copy from a remote environment.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>