HPCC-18755 Remove translated segMonitor support
Translated segmonitors are not serializable and cannot be executed remotely
(because of the associated generated code).

The value of 'offset' passed as the second parameter to createTranslatedKeySegmentMonitor
was random. This implies that the 'docompare' and 'copy' (used by index search
code will not correctly, work, though matchesBuffer (which is used when
using this SegMonitor for filtering) is ok.

Because they can't actually work, it's safest to remove them and greatly
simplifies refactoring.

VarOffset segmonitors similarly will not currently work in in-memory indexes
as the code for comparing rows is not implemented. Once refactoring of
segmonitors is completed this should become moot and they will no longer be
needed. Again, best to remove, to simplify the refactoring.

Distinction between docompare and docompareraw is now moot. All
implementations were identical, so simplify the code.

Never generate wild segmonitors - they can be added by the runtime.

Remove test that requires VarOffset segmonitors, as they are (temporarily, I
hope) unsupported.

Fixed some lurking bugs revealed by making finish() do some actual work.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>