HPCC-18784 Ensure windows share translated
HPCC-18556 ensured that URL's with a windows share were
translated to Windows format so that dafilesrv's would read/write
to the correct locations.
However, remote filenames of this type also needed to translate
the URL for the purposes of IFile->queryFilename()
This was noticed as a bug in backupnode, which generated lists
of filenames based on queryFilename() and on legacy systems that
used windows share paths, a corruption was noticed.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>