HPCC-11822 Improve archived workunit paging in WsWorkunits
1. Call getSortedDirectoryIterator() to retrieve the data about
archived workunits; 2. For archived workunits, the beforeWU/
afterWU and/or StartDate/EndDate are used to skip the directories
which store WUs not in the date range. The beforeWU/afterWU is
also used to skip the WUs which are not in the time range. 3. For
'online' workunits, the beforeWU/afterWU and/or StartDate/EndDate
are used to filter out the WUs which are not in the time range.
4. Remove PageStartFrom from paging request; 5. Refacter the code
inside WUiterate().

Also fixed 2 bugs: 1. unsorted pages are returned by the WUiterate;
2. a '-' is needed for Before/After in SashaCommand.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>