HPCC-18919 Fix potential choosen stall involving conditional.
A choosen above a conditional could stall during stop if the
conditional stopped it before it was started.
This fix prevents it trying to send the count if not started.
It also fixes an uninitialized member that was being tested
for the send, which was masking this problem most of the time.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>