HPCC-18937 Parse in child query regression, reusing last results
Introduced by HPCC-15320, the iterator was reset before the
previous results were cleared by the parser.
Each time a new input row is parsed, the parser resets the result
array in the parse, however the introduction of a parse row
iterator first() call in start(), meant that the previous results
were output again before the next row was read.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>