HPCC-18986 Fix potential stall with csvread with header lines
Csv read tried to read the header lines on start(), if the file
was sparse and the slave dealt with multiple parts per slave,
the processing might not send a message to the next slave to
indicate how many header lines had been read, until the next or
all parts had been read by that slave.
Consequently, when a csv read like this was being read by an
activity which did not start reading on all slaves, there could
be deadlock. The particular scenario this was seen in was where
as CSV read with a HEADING that was larger than the cluster, was
being read by a JOIN on the RHS.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>