HPCC-19447 Scope iterator failing if base scope given
An error 'Logical filename cannot end with scope "."' was being
hit if a base scope was provided, picked up by validation added
in HPCC-18645.

The scope iterator was incorrectly adding a trailing ".".
It was also quite inefficient because even if a deep base scope
was provided it was always connecting from the root of the files.
Changed to connect directly.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>