HPCC-19563 COMPRESSED(ROW) issues
Loading a blob from a row-compressed index would crash.

The needCopy flag was not properly passed in meaning memory-mapped indexes
would still use the node buffers.

Better to use a derived class for such things as compareRowAt, getValueAt etc
to avoid conditional tests in the normal cases (which need to be fast).

Also contains a fix for HPCC-19564 Partition keys issue.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>