HPCC-19581 Add redirect url check, new CORS check, etc.
The code is added to avoid invalid url being used to redirect.
In the existing code, a request is treated as a CORS request if
an Origin header exists. But, some request is not CORS request
even if there is an Origin header. The code is added to check
if the request comes from the same ESP service. If yes, the
request is not a CORS request. A new ESPAuthencated cookie is
added for an ESP client to check whether the client has been
authenticated or not.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>