HPCC-19755 Improve WsWorkunits.WUResultBin - file contents read
1. if a logical file is created by a DFU workunit, the existing
WUResultBin cannot read the file contents due to the failure of
retrieve cluster name though the DFU workunit. In this fix, the
code is added to retrieve cluster name based on DFU workunit id.
2. if an ESP client already knows the cluster name, I am adding
a cluster name into the WUResultBin request so that the name
can be used directly without DFU workunit.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>