HPCC-19601 Connect logMsgManager only when dali is available
In HPCC-19504, the connectLogMsgManagerToDali() call is added when
ESP is started in order to access dali audit log. That causes that
 some esp like configMgr cannot be started since the esp does not
has a dali connection. In fix moves the call into ESP's initDali()
function. Also moves the disconnectLogMsgManagerFromDali() call
to the detachESPFromDali(). A dali detach check is added into
WsDfu.DFUDeleteFiles() so that a user cannot delete a logical
file when dali is detached.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>